requirements.txt
bidict==0.23.1 ; python_version >= "3.10" and python_version < "4.0"
blinker==1.9.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
colorcet==3.1.0 ; python_version >= "3.10" and python_version < "4.0"
contourpy==1.3.2 ; python_version >= "3.10" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.10" and python_version < "4.0"
flask-socketio==5.5.1 ; python_version >= "3.10" and python_version < "4.0"
flask==3.1.0 ; python_version >= "3.10" and python_version < "4.0"
fonttools==4.57.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.16.0 ; python_version >= "3.10" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.10" and python_version < "4.0"
kiwisolver==1.4.8 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.10" and python_version < "4.0"
matplotlib==3.10.1 ; python_version >= "3.10" and python_version < "4.0"
networkx==3.4.2 ; python_version >= "3.10" and python_version < "4.0"
numpy==2.2.5 ; python_version >= "3.10" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.10" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.10" and python_version < "4.0"
pillow==11.2.1 ; python_version >= "3.10" and python_version < "4.0"
pygame==2.6.1 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.2.3 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
python-engineio==4.12.0 ; python_version >= "3.10" and python_version < "4.0"
python-socketio==5.13.0 ; python_version >= "3.10" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.10" and python_version < "4.0"
scipy==1.15.2 ; python_version >= "3.10" and python_version < "4.0"
seaborn==0.13.2 ; python_version >= "3.10" and python_version < "4.0"
simple-websocket==1.1.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.10" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.10" and python_version < "4.0"
werkzeug==3.1.3 ; python_version >= "3.10" and python_version < "4.0"
wsproto==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
shapely